remove auto increment on Faction

+++ database/migrations/2023_04_19_094348_create_faction.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::disableForeignKeyConstraints();
         Schema::dropIfExists('faction');
         Schema::create('faction', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigInteger('id');
             $table->string('name');
             $table->integer('totem')->default(0);
             $table->integer('koth')->default(0);
